refactor: export string parser as a module
